Common model code moved to approval controller
Conflicts:
	app/models/user.rb
